+++ CHANGELOG.md
@@ -6,9 +6,8 @@
 
 Some 'invalid' `Point` methods have been removed:
 
-- broadcasting `between.(pt1, pt2, range)` now needs `Ref`; `between.(Ref(pt1), Ref(pt2), range)` (or use the new for 4.0 method `between(pt1, pt2, range)`)
-
-- `Point(x, y) .+ n` is no longer valid. Use `Point(x, y) .+ Point(m, n)`.
+- Broadcasting on xy-elements like `Point(x, y) .+ n` are no longer valid. Use `Point(x, y) + Point(n, n)`.
+- `Point`-`Real` arithmetic operations such as `Point(x, y) + n` are also no longer valid.
 
 LaTeX support is still under development. 
 See https://github.com/JuliaGraphics/Cairo.jl/pull/357.